refactor: move user creation requests to module
