+++ src/modules/externalUserTypes.ts
@@ -1,0 +1,6 @@
+export interface ExternalUserUpdate {
+    email: string;
+    name?: string;
+    supervisorId?: string | null;
+    active?: boolean;
+}
